% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_minus_min.R
\name{max_minus_min}
\alias{max_minus_min}
\title{Max minus min}
\usage{
max_minus_min(x, na.rm = TRUE)
}
\arguments{
\item{x}{Group of num values}

\item{na.rm}{Input TRUE to remove NAs}
}
\value{
Max minus min num value
}
\description{
Subtract min value from max value from a group of numbers. Handle NA by removing them.
}
\examples{
max_minus_min(1:10)

}
